third_party/blink/web_tests/platform/win/fast/forms/select-popup/popup-menu-resize-after-open-expected.txt

PASS internals.pagePopupWindow.innerWidth is 100
PASS internals.pagePopupWindow.innerWidth is 150
PASS successfullyParsed is true

TEST COMPLETE

o1
o2
